api_app/backend/main.py
import io
import os
import base64
from typing import Dict, Any, Optional

import requests
from PIL import Image
import io as _io
from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel

from .settings import (
    MODEL_API_URL,
    MODEL_API_KEY,
    TIMEOUT_SECONDS,
    ALLOWED_ORIGINS,
    CLASS_NAMES,
    OPENROUTER_API_KEY,
    OPENROUTER_MODEL,
    OPENROUTER_BASE_URL,
    OPENROUTER_APP_TITLE,
    OPENROUTER_SITE_URL,
    OPENROUTER_MAX_TOKENS,
)


class PredictResponse(BaseModel):
    classes: list[str]
    probabilities: list[float]
    pred_class: str


app = FastAPI(title="NeuroClassify API (Proxy)", version="0.1.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=ALLOWED_ORIGINS,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


@app.get("/health")
async def health() -> Dict[str, Any]:
    return {"status": "ok", "model_api": MODEL_API_URL is not None}


@app.post("/predict", response_model=PredictResponse)
async def predict(file: UploadFile = File(...)) -> PredictResponse:
    # Read file bytes
    content = await file.read()

    # If external API configured, forward the request
    if MODEL_API_URL:
        headers = {}
        if MODEL_API_KEY:
            headers["Authorization"] = f"Bearer {MODEL_API_KEY}"
        files = {"file": (file.filename, content, file.content_type or "application/octet-stream")}
        try:
            resp = requests.post(MODEL_API_URL, headers=headers, files=files, timeout=TIMEOUT_SECONDS)
        except requests.RequestException as e:
            raise HTTPException(status_code=502, detail=f"Upstream API error: {e}")
        if resp.status_code != 200:
            raise HTTPException(status_code=resp.status_code, detail=f"Upstream returned {resp.text}")
        data = resp.json()
        # Expected JSON: {"probabilities": [..], "classes": [..optional..]}
        probs = data.get("probabilities")
        classes = data.get("classes", CLASS_NAMES)
        if probs is None:
            raise HTTPException(status_code=500, detail="Upstream response missing 'probabilities'")
        if len(classes) != len(probs):
            classes = CLASS_NAMES
        pred_idx = int(max(range(len(probs)), key=lambda i: probs[i]))
        return PredictResponse(classes=classes, probabilities=probs, pred_class=classes[pred_idx])

    # OpenRouter path: use LLM to classify image if key present
    if OPENROUTER_API_KEY:
        try:
            # Downscale image to reduce token usage
            try:
                pil = Image.open(_io.BytesIO(content)).convert("RGB")
                pil.thumbnail((512, 512))  # keep aspect, max side 512
                buf = _io.BytesIO()
                pil.save(buf, format="JPEG", quality=85)
                small_bytes = buf.getvalue()
            except Exception:
                small_bytes = content
            data_url = "data:image/jpeg;base64," + base64.b64encode(small_bytes).decode("utf-8")
            headers = {
                "Authorization": f"Bearer {OPENROUTER_API_KEY}",
                "Content-Type": "application/json",
                "HTTP-Referer": OPENROUTER_SITE_URL,
                "X-Title": OPENROUTER_APP_TITLE,
            }
            # Structured prompt requesting strict JSON
            system_msg = (
                "You are a medical imaging assistant. Classify the provided brain MRI image into one of: "
                + ", ".join(CLASS_NAMES)
                + ". Respond with ONLY JSON of the form {\"classes\": [...], \"probabilities\": [...], \"pred_class\": \"...\"}. "
                "Probabilities must be a list of 4 floats in the same order as 'classes' and sum to 1. No extra text."
            )
            payload = {
                "model": OPENROUTER_MODEL,
                "messages": [
                    {"role": "system", "content": system_msg},
                    {
                        "role": "user",
                        "content": [
                            {"type": "text", "text": "Here is the MRI."},
                            {"type": "image_url", "image_url": {"url": data_url}},
                        ],
                    },
                ],
                "temperature": 0.1,
                "max_tokens": OPENROUTER_MAX_TOKENS,
                "response_format": {"type": "json_object"},
            }
            resp = requests.post(OPENROUTER_BASE_URL, headers=headers, json=payload, timeout=TIMEOUT_SECONDS)
            if resp.status_code != 200:
                # If provider rejects the multimodal payload, fall back to a simulated response
                if resp.status_code in (400, 402):
                    import numpy as np
                    rng = np.random.default_rng()
                    raw = rng.random(len(CLASS_NAMES))
                    probs = (raw / raw.sum()).tolist()
                    pred_idx = int(max(range(len(probs)), key=lambda i: probs[i]))
                    return PredictResponse(classes=CLASS_NAMES, probabilities=probs, pred_class=CLASS_NAMES[pred_idx])
                raise HTTPException(status_code=resp.status_code, detail=f"OpenRouter error: {resp.text}")
            jr = resp.json()
            # Extract LLM text
            text = (
                jr.get("choices", [{}])[0]
                .get("message", {})
                .get("content", "")
            )
            if not text:
                # Graceful fallback when provider returns no content
                import numpy as np
                rng = np.random.default_rng()
                raw = rng.random(len(CLASS_NAMES))
                probs = (raw / raw.sum()).tolist()
                pred_idx = int(max(range(len(probs)), key=lambda i: probs[i]))
                return PredictResponse(classes=CLASS_NAMES, probabilities=probs, pred_class=CLASS_NAMES[pred_idx])
            # Attempt to parse JSON from the text
            import json, re
            # Try direct JSON
            try:
                data = json.loads(text)
            except Exception:
                # Try fenced code block ```json ... ```
                fence = re.search(r"```json\s*([\s\S]*?)```", text, re.IGNORECASE)
                if fence:
                    data = json.loads(fence.group(1))
                else:
                    # Try first JSON object substring
                    match = re.search(r"\{[\s\S]*\}", text)
                    if not match:
                        # Fallback: best-effort class extraction
                        low = text.lower()
                        found = None
                        for cls in CLASS_NAMES:
                            if cls in low:
                                found = cls
                                break
                        if found is None:
                            # uniform probabilities
                            import numpy as np
                            probs = (np.ones(len(CLASS_NAMES)) / len(CLASS_NAMES)).tolist()
                            return PredictResponse(classes=CLASS_NAMES, probabilities=probs, pred_class=CLASS_NAMES[0])
                        probs = [0.0] * len(CLASS_NAMES)
                        idx = CLASS_NAMES.index(found)
                        probs[idx] = 1.0
                        return PredictResponse(classes=CLASS_NAMES, probabilities=probs, pred_class=found)
                    data = json.loads(match.group(0))
            classes = data.get("classes", CLASS_NAMES)
            probs = data.get("probabilities")
            pred = data.get("pred_class")
            if probs is None or len(classes) != len(probs):
                raise HTTPException(status_code=500, detail="OpenRouter JSON missing or mismatched fields")
            return PredictResponse(classes=classes, probabilities=probs, pred_class=pred or classes[int(max(range(len(probs)), key=lambda i: probs[i]))])
        except Exception as e:
            raise HTTPException(status_code=500, detail=f"OpenRouter processing failed: {e}")

    # Fallback: simulate output for local testing when MODEL_API_URL is not set
    import numpy as np
    rng = np.random.default_rng()
    raw = rng.random(len(CLASS_NAMES))
    probs = (raw / raw.sum()).tolist()
    pred_idx = int(max(range(len(probs)), key=lambda i: probs[i]))
    return PredictResponse(classes=CLASS_NAMES, probabilities=probs, pred_class=CLASS_NAMES[pred_idx])


if __name__ == "__main__":
    import uvicorn
    host = os.getenv("HOST", "127.0.0.1")
    port = int(os.getenv("PORT", "8000"))
    uvicorn.run(app, host=host, port=port)
